fix some warnings
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@3247 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/pdf/pdfobj.c
@@ -31,7 +31,7 @@ integer pdf_last_obj;
 void pdf_write_obj(PDF pdf, integer k)
 {
     lstring data, st;
-    integer i;                  /* index into |data.s| */
+    size_t i;                   /* index into |data.s| */
     int saved_compress_level = pdf->compress_level;
     int os_level = 1;           /* gives compressed objects for \pdfobjcompresslevel > 0 */
     int l = 0;                  /* possibly a lua registry reference */
